+++ changelog.md
@@ -47,6 +47,7 @@
 ### Updated Libraries
 - jQuery Mobile to 1.4.5
 - jQuery to 2.1.4
+- Highcharts to 5.0.14, migrated to styled mode
 
 ### Fixed Bugs
 - Changes were not visible immediate after saving configuration